BUG: Simplify and fix most logic using cut model
The original roi generated with the principal axes of the input bone
model is now always saved in the TreeNode `self.roi` class field. In
each frame, we assess whether that roi exceeds the target volume, and if
so we generate a cropped model and corresponding roi for that frame.
The new roi is used just for cropping that target frame volume and it's
not saved for later use. This way, in the next frame, the original roi
is still saved and we'll compare it again to the bounds of the next
frame fov (so no need to recompute it).

NOTE: I am not sure that generating the roi from the cut model currently
works as expected, I observed the result ROI looking the same as the
oriented ROI of the original model. However, as of this commit, the
model and roi are properly spacially aligned in order to generate the
expected cropped source and target volumes in each frame.