BUG: Generate model roi from principle axis and use logic in Out of bounds clip  (#185)
* BUG: Generate roi in model-defined orientation

Generated using segmentation statistics module, each model is cloned and a principle rotation axis computed to form model-derived coordinate system basis.

Roi are constructed from bounding box dimensions of the models in the space and back transformed to source space by the inverse linear transform.

* STYLE: Update Hierarchical3DRegistration/Hierarchical3DRegistrationLib/TreeNode.py

Co-authored-by: Shelly Belsky <71195502+sbelsk@users.noreply.github.com>

* ENH: Investigate better roi adjustment for out of bounds frames

* BUG: Simplify and fix most logic using cut model

The original roi generated with the principal axes of the input bone
model is now always saved in the TreeNode `self.roi` class field. In
each frame, we assess whether that roi exceeds the target volume, and if
so we generate a cropped model and corresponding roi for that frame.
The new roi is used just for cropping that target frame volume and it's
not saved for later use. This way, in the next frame, the original roi
is still saved and we'll compare it again to the bounds of the next
frame fov (so no need to recompute it).

NOTE: I am not sure that generating the roi from the cut model currently
works as expected, I observed the result ROI looking the same as the
oriented ROI of the original model. However, as of this commit, the
model and roi are properly spacially aligned in order to generate the
expected cropped source and target volumes in each frame.

* ENH: Hide in development scene nodes

Also ensure pre-commit criteria passed

* BUG: Return appropriate positive or negative model

Previously only export positive model from plane modeler. Based on index of boundary extension main or max returns appropriate model.

* BUG: Positive Negative model flip with lpsras

WIP TO DO: remove commented out remove node from scene lines once tested

* STYLE: Remove debug commented node removals

---------

Co-authored-by: Shelly Belsky <71195502+sbelsk@users.noreply.github.com>
Co-authored-by: Shelly Belsky <shelly.belsky@kitware.com>